Create `mosaic-testing` module
