+++ mosaic-runtime/build.gradle
@@ -31,6 +31,7 @@ kotlin {
 		}
 		commonTest {
 			dependencies {
+				implementation projects.mosaicTesting
 				implementation libs.kotlin.test
 				implementation libs.kotlinx.coroutines.test
 				implementation libs.assertk
