+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/layout/AspectRatioTest.kt
@@ -7,9 +7,9 @@ import com.jakewharton.mosaic.NodeSnapshots
 import com.jakewharton.mosaic.TestFiller
 import com.jakewharton.mosaic.assertFailure
 import com.jakewharton.mosaic.modifier.Modifier
-import com.jakewharton.mosaic.runMosaicTest
 import com.jakewharton.mosaic.size
 import com.jakewharton.mosaic.testIntrinsics
+import com.jakewharton.mosaic.testing.runMosaicTest
 import com.jakewharton.mosaic.ui.Layout
 import com.jakewharton.mosaic.ui.unit.Constraints
 import com.jakewharton.mosaic.ui.unit.IntSize
