+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/layout/SizeTest.kt
@@ -8,9 +8,9 @@ import com.jakewharton.mosaic.Container
 import com.jakewharton.mosaic.NodeSnapshots
 import com.jakewharton.mosaic.modifier.Modifier
 import com.jakewharton.mosaic.position
-import com.jakewharton.mosaic.runMosaicTest
 import com.jakewharton.mosaic.size
 import com.jakewharton.mosaic.testIntrinsics
+import com.jakewharton.mosaic.testing.runMosaicTest
 import com.jakewharton.mosaic.ui.Alignment
 import com.jakewharton.mosaic.ui.Box
 import com.jakewharton.mosaic.ui.Column
