+++ mosaic-testing/api/mosaic-testing.api
@@ -1,0 +1,14 @@
+public abstract interface class com/jakewharton/mosaic/testing/SnapshotStrategy {
+	public abstract fun create (Lcom/jakewharton/mosaic/Mosaic;)Ljava/lang/Object;
+}
+
+public abstract interface class com/jakewharton/mosaic/testing/TestMosaic : com/jakewharton/mosaic/Mosaic {
+	public abstract fun awaitSnapshot-VtjQ1oo (JLkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public static synthetic fun awaitSnapshot-VtjQ1oo$default (Lcom/jakewharton/mosaic/testing/TestMosaic;JLkotlin/coroutines/Continuation;ILjava/lang/Object;)Ljava/lang/Object;
+}
+
+public final class com/jakewharton/mosaic/testing/TestMosaicKt {
+	public static final fun runMosaicTest (Lcom/jakewharton/mosaic/testing/SnapshotStrategy;Lkotlin/jvm/functions/Function2;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public static final fun runMosaicTest (Lkotlin/jvm/functions/Function2;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
