+++ mosaic-testing/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=mosaic-testing
+POM_NAME=Mosaic testing
