+++ settings.gradle
@@ -19,6 +19,7 @@ rootProject.name = 'mosaic'
 include ':mosaic-animation'
 include ':mosaic-runtime'
 include ':mosaic-terminal'
+include ':mosaic-testing'
 
 include ':samples:counter'
 include ':samples:demo'
